Fix broken specs
